requirements.txt
streamlit>=1.24.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
ultralytics>=8.0.0
pandas>=2.0.0
tensorflow>=2.13.0
tensorflow-hub>=0.14.0 